+++ src/gaos/models/interactions/interaction-completed-event.ts
@@ -12,6 +12,13 @@
 
 import { InteractionSseEventInteraction } from "./interaction-sse-event-interaction.js";
 
+/**
+ * Signals that the Interaction completed. Sent when the Interaction receives
+ *
+ * @remarks
+ * Complete/Cancel or naturally terminates. No more input can be sent to the
+ * Interaction after this.
+ */
 export type InteractionCompletedEvent = {
   /**
    * The event_id token to be used to resume the interaction stream, from