+++ src/gaos/models/operations/create-agent.ts
@@ -14,18 +14,18 @@ import * as agents from "../agents/index.js";
 
 export type CreateAgentGlobals = {
   /**
-   * Which version of the API to use.
+   * API version for request routing.
    */
   api_version?: string | undefined;
 };
 
 export type CreateAgentRequest = {
   /**
-   * Which version of the API to use.
+   * API version for request routing.
    */
   api_version?: string | undefined;
   /**
-   * The request body.
+   * Required. The request body.
    */
   body: agents.Agent;
 };